+++ packages/tools-i18n/.npmignore
