+++ packages/tools-i18n/.npmrc
