+++ packages/tools-i18n/lib/defaults.js
