+++ packages/tools-i18n/lib/json-imports.js
