+++ packages/tools-i18n/lib/toJSON.js
