+++ packages/tools-i18n/package.json
@@ -1,6 +1,6 @@
 {
-  "name": "@ui5/webcomponents-i18n",
-  "version": "1.0.0-rc.12",
+  "name": "@ui5/webcomponents-tools-i18n",
+  "version": "1.0.0-rc.14",
   "description": "UI5 Web Components: webcomponents.i18n",
   "author": "SAP SE (https://www.sap.com)",
   "license": "Apache-2.0",
