+++ packages/tools/lib/generate-json-imports/i18n.js
@@ -1,1 +1,1 @@
-require("@ui5/webcomponents-i18n/lib/json-imports.js");
+require("@ui5/webcomponents-tools-i18n/lib/json-imports.js");
