+++ packages/tools/lib/i18n/toJSON.js
@@ -1,1 +1,1 @@
-require("@ui5/webcomponents-i18n/lib/toJSON.js");
+require("@ui5/webcomponents-tools-i18n/lib/toJSON.js");
